\name{calc.price.adjust}
\alias{calc.price.adjust}
\title{adjust futures price data}
\description{
calculate adjustment for backwards price adjusted data
}
\usage{
calc.price.adjust(contract.list)
}
\arguments{
  \item{contract.list}{a list of individual futures contracts}
}
\value{
a perpetual futures contract
}
\references{LIM}
\author{Whit}
\keyword{ts}

